pyproject.toml
[tool.poetry]
name = "mori"
version = "0.1.0"
description = "Project with Daryl & Quentin for the rncp"
authors = ["Santos <santoslukombo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
pytest = "^8.3.3"
uvicorn = "^0.32.0"
pytest-asyncio = "^0.24.0"
pydotenv = "^0.0.7"
python-dotenv = "^1.0.1"
ollama = "^0.4.8"
python-jose = "^3.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"